Add UI to visualize and configure locales

+++ MobileTracking.Core/Commands/CreateOrUpdateLocaleCommand.cs
@@ -8,8 +8,8 @@ namespace MobileTracking.Core.Application
 
         public string? Description { get; set; }
 
-        public float? Latitude { get; set; }
+        public double? Latitude { get; set; }
 
-        public float? Longitude { get; set; }
+        public double? Longitude { get; set; }
     }
 }
